backend complete

+++ STUDYNOTION/SERVER/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require('../models/Section')
 const Course = require('../models/Course')
+const SubSection = require('../models/SubSection')
 
 //create Section
 exports.createSection = async (req, res) => {
@@ -20,27 +21,27 @@ exports.createSection = async (req, res) => {
 
         //update course with section objectId
         const updatedCourseDetails = await Course.findByIdAndUpdate(
-                                            courseId,
-                                            {
-                                                $push: {
-                                                    courseContent: newSection._id
-                                                }
-                                            },
-                                            { new: true }
-                                        )
-                                        .populate({
-                                            path:"courseContent",
-                                            populate: {
-                                                path: "SubSection",
-                                            },
-                                        })
-                                        .exec()
+            courseId,
+            {
+                $push: {
+                    courseContent: newSection._id
+                }
+            },
+            { new: true }
+        )
+            .populate({
+                path: "courseContent",
+                populate: {
+                    path: "subsection",
+                },
+            })
+            .exec()
         //HW: use populate to replace section/sub-section both in the updatedcourse
 
         //return response
         return res.status(200).json({
-            success:true,
-            message:"section created successfully",
+            success: true,
+            message: "section created successfully",
             updatedCourseDetails
         })
     }
@@ -56,10 +57,10 @@ exports.createSection = async (req, res) => {
 
 
 // update Section
-exports.updateSection = async (req,res) =>{
-    try{
+exports.updateSection = async (req, res) => {
+    try {
         //fetch data
-        const {sectionName, sectionId} = req.res
+        const { sectionName, sectionId } = req.body
 
         //validate
         if (!sectionName || !sectionId) {
@@ -70,13 +71,13 @@ exports.updateSection = async (req,res) =>{
         }
 
         //update section
-        const updatedSection = await Section.findByIdAndUpdate(sectionId,{sectionName},{new:true})
+        const updatedSection = await Section.findByIdAndUpdate(sectionId, { sectionName }, { new: true })
 
         //return response
         return res.status(200).json({
-            success:true,
-            message:"section updated successfully",
-            updatedCourseDetails
+            success: true,
+            message: "section updated successfully",
+            updatedSection
         })
     }
     catch (error) {
@@ -91,20 +92,58 @@ exports.updateSection = async (req,res) =>{
 
 
 // delete Section
-exports.deleteSection = async (req,res)=>{
-    try{
+exports.deleteSection = async (req, res) => {
+    try {
         //get Id :- assuming that we are sending Id in params
-        const {sectionId}=req.params
+        const { sectionId, courseId } = req.body
 
-        // findByIdandDelete
-        await Section.findByIdAndDelete(sectionId)
-        //Todo:- do we need to delete the entry from the course schema
+        // Validate input
+        if (!sectionId || !courseId) {
+            return res.status(400).json({
+                success: false,
+                message: "Both sectionId and courseId are required",
+            });
+        }
+
+        //delete section from course
+        const updatedCourse = await Course.findByIdAndUpdate(
+            courseId,
+            {
+                $pull: {
+                    courseContent: sectionId
+                }
+            }
+        )
+
+        if (!updatedCourse) {
+            return res.status(404).json({
+                success: false,
+                message: "Course not Found",
+            })
+        }
+
+        // find section and delete
+        const deletedSection = await Section.findByIdAndDelete(sectionId)
+        if (!updatedSection) {
+            return res.status(404).json({
+                success: false,
+                message: "Section not Found",
+            })
+        }
+
+        //delete all the subsection 
+        await SubSection.deleteMany({
+            _id:
+                { $in: updatedSection.subsection }
+        })
+
 
         //return response
         return res.status(200).json({
-            success:true,
-            message:"section deleted successfully",
-            updatedCourseDetails
+            success: true,
+            message: "section deleted successfully",
+            updatedSection
+            // updatedCourseDetails
         })
     }
     catch (error) {
